Make the filesystem use size_t for read/write size and cast down before calling the OS function. Removes the need for me to cast all the time, and the error checking can just be in one place.

+++ Pulp/3DEngine/XModel.cpp
@@ -497,7 +497,7 @@ bool XModel::LoadModel(core::XFile* file)
 	// which we can load all at once and just set pointers.
 	//	char* pData = X_NEW_ARRAY_ALIGNED(char, hdr_.dataSize, g_3dEngineArena, "ModelBuffer", 8);
 	char* pData = X_NEW_ARRAY_ALIGNED(char, hdr_.dataSize, gEnv->pArena, "ModelBuffer", 8);
-	uint32_t readSize;
+	size_t readSize;
 
 	if ((readSize = file->read(pData, hdr_.dataSize)) != hdr_.dataSize)
 	{
